# Optimize some operations of DiscreteTrajectory and friends